requirements.txt
streamlit==1.37.0
pillow==10.4.0
tensorflow==2.17.0
keras==3.6.0
numpy==1.26.3